modified:   Makefile         modified:   checkpoints/phase7/training_log.json         new file:   configs/phase8_10b_ultra.yaml         modified:   configs/phase8_max_push.yaml         new file:   configs/phase8_max_push_300m.yaml         new file:   configs/phase8_max_push_ultra_compressed.yaml         new file:   debug_bitnet.py         new file:   debug_env.py         new file:   docs/SPEED_OPTIMIZATION_GUIDE.md         new file:   docs/TRAINING_ACCELERATION_ROADMAP.md         modified:   requirements.txt         new file:   results/benchmarks/HTT_COMPRESSION_COMPARISON_JP.md         new file:   results/benchmarks/PHASE7_VS_PHASE8_TRITON_COMPARISON_JP.md         new file:   results/benchmarks/PHASE8_MAX_PARAMS_VERIFICATION_JP.md         new file:   results/benchmarks/PHASE8_TRITON_WSL_BENCHMARK_EN.md         new file:   results/benchmarks/PHASE8_TRITON_WSL_BENCHMARK_JP.md         new file:   results/benchmarks/htt_compression_comparison.json         new file:   results/benchmarks/phase8_comprehensive_benchmark.json         new file:   results/benchmarks/phase8_triton_wsl_results.json         new file:   results/speed_benchmark.txt         new file:   rust_loader/Cargo.toml         new file:   rust_loader/pyproject.toml         new file:   rust_loader/src/lib.rs         modified:   scripts/benchmark_consumer_gpu.py         new file:   scripts/benchmark_htt_compression_comparison.py         new file:   scripts/benchmark_phase8_comprehensive.py         modified:   scripts/benchmark_phase8_memory.py         new file:   scripts/benchmark_phase8_speed.py         new file:   scripts/benchmark_phase8_triton_quick.py         modified:   scripts/benchmark_quantization.py         new file:   scripts/calc_10b_config.py         new file:   scripts/check_phase8_max_params.py         new file:   scripts/check_phase8_params_wsl.py         new file:   scripts/estimate_phase8_params.py         new file:   scripts/quick_param_check.py         new file:   scripts/run_phase8_benchmark_wsl.sh         new file:   scripts/run_phase8_param_check.sh         new file:   scripts/run_phase8_param_check_wsl.bat         new file:   scripts/run_phase8_param_check_wsl.sh         modified:   scripts/train_phase7.py         modified:   scripts/train_phase8.py         new file:   scripts/verify_phase8_max_params.py         new file:   src/benchmarks/benchmark_inference.py         new file:   src/benchmarks/benchmark_loader.py         new file:   src/benchmarks/optimization_benchmark.py         new file:   src/inference/inference_engine.py         new file:   src/kernels/bitnet_triton.py         modified:   src/kernels/bk_scan.py         new file:   src/kernels/flash_attention_wrapper.py         new file:   src/kernels/htt_triton.py         modified:   src/kernels/quantized_hyperbolic_triton.py         new file:   src/models/bitnet.py         modified:   src/models/bk_core.py         modified:   src/models/config.py         modified:   src/models/phase1/htt_embedding.py         modified:   src/models/phase7/hybrid_attention.py         modified:   src/models/phase7/hyperbolic_attention.py         modified:   src/models/phase7/integrated_model.py         modified:   src/models/phase8/ar_ssm_fusion.py         modified:   src/models/phase8/bk_core_hyperbolic.py         modified:   src/models/phase8/config.py         modified:   src/models/phase8/hyperbolic_ssm.py         modified:   src/models/phase8/linear_attention.py         modified:   src/models/resnet_bk.py         modified:   src/models/semiseparable_matrix.py         modified:   src/utils/data_utils.py         new file:   src/utils/dataloader_optimizations.py         new file:   tests/test_bitnet_triton.py         new file:   tests/test_bk_scan_fused.py         new file:   tests/test_htt_triton.py         new file:   tests/test_inference_engine.py         new file:   u00261

+++ src/models/phase8/ar_ssm_fusion.py
@@ -282,7 +282,7 @@ class AdaptiveRankSSM(nn.Module):
         
         diagnostics = {
             'state_norm_mean': h.norm(dim=-1).mean(),
-            'A_spectral_norm': torch.linalg.matrix_norm(A, ord=2),
+            'A_spectral_norm': torch.linalg.matrix_norm(A.float(), ord=2),
         }
         
         return output, diagnostics
